# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "EDCOCR API",
    "description": "REST API for forensic-grade document OCR processing.",
    "version": "4.1.0"
  },
  "paths": {
...
    "/api/v1/search/semantic": {
      "post": {
        "tags": [
          "semantic"
        ],
        "summary": "Semantic Search",
        "description": "Search documents using semantic understanding via VLM.\n\nSends the query to the configured VLM inference server and returns\nranked results with relevance scores and bounding boxes.",
        "operationId": "semantic_search_api_v1_search_semantic_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SemanticSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SemanticSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/search/analyze": {
      "post": {
        "tags": [
          "semantic"
        ],
        "summary": "Analyze Document",
        "description": "Analyze document pages using a VLM for entity extraction and summarization.\n\nSends pages (text and optional images) to the VLM inference server\nfor deep document understanding.",
        "operationId": "analyze_document_api_v1_search_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentAnalysisResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "502": {
            "description": "Bad Gateway",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "504": {
            "description": "Gateway Timeout",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/api__models__ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/entities": {
      "get": {
        "tags": [
          "recall"
        ],
        "summary": "Search Entities",
        "description": "Search indexed entities across all processed documents.\n\nSupports filtering by entity type, text content, job ID, and\nminimum confidence. Results are ordered by most recently indexed.",
        "operationId": "search_entities_api_v1_entities_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by entity type (e.g. PERSON, DATE)",
              "title": "Type"
            },
            "description": "Filter by entity type (e.g. PERSON, DATE)"
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Text search query (LIKE matching)",
              "title": "Q"
            },
            "description": "Text search query (LIKE matching)"
          },
          {
            "name": "job_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by job ID",
              "title": "Job Id"
            },
            "description": "Filter by job ID"
          },
          {
            "name": "min_confidence",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 1.0,
              "minimum": 0.0,
              "description": "Minimum confidence threshold",
              "default": 0.0,
              "title": "Min Confidence"
            },
            "description": "Minimum confidence threshold"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Page size",
              "default": 50,
              "title": "Limit"
            },
            "description": "Page size"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of items to skip",
              "default": 0,
              "title": "Offset"
            },
            "description": "Number of items to skip"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitySearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentAnalysisRequest": {
        "properties": {
          "pages": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "minItems": 1,
            "title": "Pages",
            "description": "List of page dicts with page_number, text, and optional image_b64."
          },
          "prompt": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4000
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt",
            "description": "Analysis instruction or prompt."
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id",
            "description": "Optional document identifier."
          }
        },
        "type": "object",
        "required": [
          "pages"
        ],
        "title": "DocumentAnalysisRequest",
        "description": "Request body for VLM document analysis."
      },
      "DocumentAnalysisResponse": {
        "properties": {
          "entities": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Entities"
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "relationships": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Relationships"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 0.0
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "DocumentAnalysisResponse",
        "description": "Response from VLM document analysis."
      },
      "EntitySearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/EntitySearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "results",
          "total",
          "limit",
          "offset"
        ],
        "title": "EntitySearchResponse",
        "description": "Paginated entity search results."
      },
      "EntitySearchResult": {
        "properties": {
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "document_name": {
            "type": "string",
            "title": "Document Name"
          }
        },
        "type": "object",
        "required": [
          "entity_id",
          "job_id",
          "entity_type",
          "text",
          "confidence",
          "source",
          "page",
          "document_name"
        ],
        "title": "EntitySearchResult",
        "description": "Single entity search result."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SemanticSearchRequest": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "description": "Natural-language search query."
          },
          "document_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id",
            "description": "Optional document ID to scope the search."
          },
          "max_results": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Max Results",
            "description": "Maximum number of results to return.",
            "default": 10
          },
          "min_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Min Score",
            "description": "Minimum relevance score threshold (0.0-1.0).",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SemanticSearchRequest",
        "description": "Request body for semantic search."
      },
      "SemanticSearchResponse": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SemanticSearchResult"
            },
            "type": "array",
            "title": "Results"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "processing_time_ms": {
            "type": "number",
            "title": "Processing Time Ms",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "query",
          "results",
          "total"
        ],
        "title": "SemanticSearchResponse",
        "description": "Response containing semantic search results."
      },
      "SemanticSearchResult": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text",
            "description": "Matched text passage."
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Relevance score (0.0-1.0)."
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Page number where the match was found."
          },
          "document_id": {
            "type": "string",
            "title": "Document Id",
            "description": "Source document identifier.",
            "default": ""
          },
          "bbox": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Bbox",
            "description": "Bounding box [x1, y1, x2, y2] of the matched region."
          }
        },
        "type": "object",
        "required": [
          "text",
          "score",
          "page"
        ],
        "title": "SemanticSearchResult",
        "description": "A single semantic search result."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "api__models__ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authentication. Set via OCR_API_KEY environment variable."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "OAuth2 Bearer token (when OAUTH2_ENABLED=true)."
      }
    }
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    },
    {
      "BearerAuth": []
    }
  ]
}